refactor: migration and recombination
