+++ src/components/CodemirrorEditor/EditorHeader/index.vue
@@ -3,13 +3,13 @@ import { nextTick, ref } from 'vue'
 import { storeToRefs } from 'pinia'
 import { ElNotification } from 'element-plus'
 
-import StyleOptionMenu from './StyleOptionMenu.vue'
 import PostInfo from './PostInfo.vue'
-import HelpDropdown from './HelpDropdown.vue'
 import FileDropdown from './FileDropdown.vue'
+import HelpDropdown from './HelpDropdown.vue'
+import StyleDropdown from './StyleDropdown.vue'
+import EditDropdown from './EditDropdown.vue'
 
-import { mergeCss, solveWeChatImage } from '@/assets/scripts/converter'
-import config from '@/config'
+import { mergeCss, solveWeChatImage } from '@/utils'
 import { useStore } from '@/stores'
 
 const emit = defineEmits([
@@ -17,9 +17,6 @@ const emit = defineEmits([
   `formatContent`,
   `startCopy`,
   `endCopy`,
-  `showCssEditor`,
-  `showDialogUploadImg`,
-  `showDialogForm`,
 ])
 
 const formatItems = [
@@ -53,37 +50,16 @@ const formatItems = [
 const store = useStore()
 
 const {
-  fontFamily,
-  fontSize,
-  fontColor,
-  codeBlockTheme,
-  legend,
   isDark,
   isCiteStatus,
-  isMacCodeBlock,
   output,
-  cssEditor,
 } = storeToRefs(store)
 
 const {
   toggleDark,
-  resetStyleConfirm,
   editorRefresh,
-
-  fontChanged,
-  sizeChanged,
-  colorChanged,
-  codeBlockThemeChanged,
-  legendChanged,
-  macCodeBlockChanged,
   citeStatusChanged,
 } = store
-
-const colorPicker = ref(null)
-
-function showPicker() {
-  colorPicker.value.show()
-}
 
 // 复制到微信公众号
 function copy() {
@@ -157,20 +133,6 @@ function copy() {
     emit(`endCopy`)
   }, 350)
 }
-
-// 自定义CSS样式
-function customStyle() {
-  emit(`showCssEditor`)
-  nextTick(() => {
-    if (!cssEditor.value) {
-      cssEditor.value.refresh()
-    }
-  })
-  setTimeout(() => {
-    cssEditor.value.refresh()
-  }, 50)
-}
-
 </script>
 
 <template>
@@ -206,104 +168,8 @@ function customStyle() {
           </el-dropdown-menu>
         </template>
       </el-dropdown>
-      <el-dropdown>
-        <span class="el-dropdown-link">
-          编辑<el-icon class="ml-2">
-            <ElIconArrowDown />
-          </el-icon>
-        </span>
-        <template #dropdown>
-          <el-dropdown-menu>
-            <el-dropdown-item class="leading-8" @click="$emit('showDialogUploadImg')">
-              <el-icon>
-                <ElIconUpload />
-              </el-icon>
-              上传图片
-            </el-dropdown-item>
-            <el-dropdown-item class="leading-8" @click="$emit('showDialogForm')">
-              <el-icon>
-                <ElIconGrid />
-              </el-icon>
-              插入表格
-            </el-dropdown-item>
-          </el-dropdown-menu>
-        </template>
-      </el-dropdown>
-      <el-dropdown>
-        <span class="el-dropdown-link">
-          样式<el-icon class="ml-2">
-            <ElIconArrowDown />
-          </el-icon>
-        </span>
-        <template #dropdown>
-          <el-dropdown-menu>
-            <el-dropdown-item class="leading-8">
-              <StyleOptionMenu title="字体" :options="config.builtinFonts" :current="fontFamily" :change="fontChanged" />
-            </el-dropdown-item>
-            <el-dropdown-item class="leading-8">
-              <StyleOptionMenu title="字号" :options="config.sizeOption" :current="fontSize" :change="sizeChanged" />
-            </el-dropdown-item>
-            <el-dropdown-item class="leading-8">
-              <StyleOptionMenu
-                title="颜色"
-                :options="config.colorOption"
-                :current="fontColor"
-                :change="colorChanged"
-              />
-            </el-dropdown-item>
-            <el-dropdown-item class="leading-8">
-              <StyleOptionMenu
-                title="代码主题"
-                :options="config.codeThemeOption"
-                :current="codeBlockTheme"
-                :change="codeBlockThemeChanged"
-              />
-            </el-dropdown-item>
-            <el-dropdown-item class="leading-8">
-              <StyleOptionMenu title="图注格式" :options="config.legendOption" :current="legend" :change="legendChanged" />
-            </el-dropdown-item>
-            <el-dropdown-item divided class="leading-8" @click="showPicker()">
-              <el-tooltip placement="right" effect="light">
-                <template #content>
-                  <el-color-picker
-                    ref="colorPicker"
-                    v-model="fontColor"
-                    :teleported="false"
-                    show-alpha
-                    class="ml-auto"
-                    style="height: 2em"
-                    @change="colorChanged"
-                  />
-                </template>
-                <div class="w-full flex">
-                  <el-icon class="opacity-0">
-                    <ElIconCheck />
-                  </el-icon>
-                  自定义颜色
-                </div>
-              </el-tooltip>
-            </el-dropdown-item>
-            <el-dropdown-item class="leading-8" @click="customStyle">
-              <el-icon class="opacity-0">
-                <ElIconCheck />
-              </el-icon>
-              自定义 CSS
-            </el-dropdown-item>
-            <el-dropdown-item divided class="leading-8" @click="macCodeBlockChanged">
-              <el-icon :class="{ 'opacity-0': !isMacCodeBlock }">
-                <ElIconCheck />
-              </el-icon>
-              Mac 代码块
-            </el-dropdown-item>
-            <el-dropdown-item divided class="leading-8" @click="resetStyleConfirm">
-              <el-icon class="opacity-0">
-                <ElIconCheck />
-              </el-icon>
-              重置
-            </el-dropdown-item>
-          </el-dropdown-menu>
-        </template>
-      </el-dropdown>
+      <EditDropdown />
+      <StyleDropdown />
       <HelpDropdown />
     </el-space>
     <el-button plain type="primary" @click="copy">
