+++ src/main.js
@@ -8,8 +8,8 @@ import 'codemirror/lib/codemirror.css'
 import 'codemirror/theme/xq-light.css'
 
 /* 每个页面公共css */
-import './assets/less/theme.less'
-import './assets/less/style-mirror.less'
+import '@/assets/less/theme.less'
+import '@/assets/less/style-mirror.less'
 
 import 'codemirror/mode/css/css'
 import 'codemirror/mode/markdown/markdown'
